+++ core-internal/src/main/java/io/github/ackeecz/guardian/core/internal/SynchronizedDataHolder.kt
@@ -24,7 +24,7 @@ public abstract class SynchronizedDataHolder<T : Any> {
             if (synchronizedData == null) {
                 synchronizedData = createSynchronizedData()
             }
-            requireNotNull(synchronizedData)
+            checkNotNull(synchronizedData)
         }
     }