+++ jetpack/src/test/java/io/github/ackeecz/guardian/jetpack/prefs/EncryptionDecryptionTest.kt
@@ -371,7 +371,7 @@ internal class EncryptionDecryptionTest : EncryptedSharedPreferencesTest() {
                 prefsFileName = prefsFileName,
             )
             decryptedPreferenceKey == preferenceKey
-        }.let(::requireNotNull)
+        }.let(::checkNotNull)
     }
 
     private fun getAndDecryptPreferenceValue(